Create SQL - Music Industry Financial Crime Investigation
For a long time, Adams Andrew, Manager of WSDA Music, has been unable to 
account for a discrepancy in his company’s financials.
The furthest he has gotten in his own attempts to analyze the company data is figuring out that 
the discrepancy occurred between the years 2011 and 2012. 
knows for certain.